refactor: renamed components, improved UI ordering
