+++ src/components/ModalRouter.tsx
@@ -7,6 +7,8 @@ import ViewAllExplorer from '../views/ViewAllExplorer';
 import { RouterCtrl } from '../controllers/RouterCtrl';
 import InitialExplorer from '../views/InitialExplorer';
 import ConnectingView from '../views/ConnectingView';
+import useTheme from '../hooks/useTheme';
+import { StyleSheet, View } from 'react-native';
 
 interface Props {
   onCopyClipboard?: (value: string) => void;
@@ -14,6 +16,7 @@ interface Props {
 
 export function ModalRouter(props: Props) {
   const routerState = useSnapshot(RouterCtrl.state);
+  const Theme = useTheme();
   const { height, width, isPortrait } = useOrientation();
 
   const ViewComponent = useMemo(() => {
@@ -32,11 +35,20 @@ export function ModalRouter(props: Props) {
   }, [routerState.view]);
 
   return (
-    <ViewComponent
-      windowHeight={height}
-      windowWidth={width}
-      isPortrait={isPortrait}
-      {...props}
-    />
+    <View style={[styles.wrapper, { backgroundColor: Theme.background1 }]}>
+      <ViewComponent
+        windowHeight={height}
+        windowWidth={width}
+        isPortrait={isPortrait}
+        {...props}
+      />
+    </View>
   );
 }
+
+const styles = StyleSheet.create({
+  wrapper: {
+    borderTopLeftRadius: 30,
+    borderTopRightRadius: 30,
+  },
+});
